Add function to update cron entry schedule
Previously, to modify the schedule of a cron entry, the entry was
deleted and a new entry with the updated schedule was put in, resulting
in a new cron task id. New cron task id must be then replaced in all
table entries that referenced the old id. Not good.

    - update_cron_schedule() updates the schedule in-place
    without any unintended consequences.

Signed-off-by: six-k <ramsri.hp@gmail.com>